Refactor filler plugin for easier maintenance
